+++ packages/flutter/lib/rendering.dart
@@ -36,6 +36,7 @@ export 'src/rendering/binding.dart';
 export 'src/rendering/box.dart';
 export 'src/rendering/custom_layout.dart';
 export 'src/rendering/debug.dart';
+export 'src/rendering/debug_overflow_indicator.dart';
 export 'src/rendering/editable.dart';
 export 'src/rendering/error.dart';
 export 'src/rendering/flex.dart';
