--- packages/flutter/lib/src/painting/debug_overflow_indicator.dart
@@ -1,392 +1,0 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'dart:math' as math;
-import 'dart:ui' as ui;
-
-import 'package:flutter/rendering.dart';
-
-// Data used by the DebugOverflowIndicator to manage the regions and labels for
-// the indicators.
-class _OverflowRegionData {
-  _OverflowRegionData({
-    this.rect,
-    this.label = '',
-    this.labelOffset = Offset.zero,
-    this.rotation = 0.0,
-  });
-
-  final Rect rect;
-  final String label;
-  final Offset labelOffset;
-  final double rotation;
-}
-
-/// An mixin indicator that is drawn when a [RenderObject] overflows its
-/// container.
-///
-/// This is used by RenderObjects that are containers to show where, and by how
-/// much, their children overflow their containers.  These indicators are
-/// typically only shown in a debug build (where the call to
-/// [overflowPaintIndicator] is surrounded by an assert).
-///
-/// Typical usage is to set the [overflowContainerRect] and the
-/// [overflowChildRect], and then call [overflowPaintIndicator]. This will
-/// calculate the overflow, and then draw the relevant indicators.
-///
-/// This class will also print a debug message to the console when the container
-/// overflows, but only on the first occurrence.  To reset and print it again
-/// upon the next occurrence, set [overflowReportNeeded] to true. This is
-/// typically done in the [RenderObject.reassemble] override in the
-/// [RenderObject] that uses the DebugOverflowIndicatorMixin.
-///
-/// ## Sample code
-///
-/// ```dart
-/// class MyRenderObject extends RenderObject with DebugOverflowContainerMixin {
-///   @override
-///   TextDirection get textDirection => _textDirection;
-///
-///   @override
-///   void performLayout() {
-///     // ...
-///     overflowChildRect = childParentData.offset & child.size;
-///     overflowContainerRect = Offset.zero & size;
-///   }
-///
-///   @override
-///   void paint(PaintingContext context, Offset offset) {
-///     if (isOverflowing) {
-///       _paintStuff(context, offset);
-///       return;
-///     }
-///     // ...
-///     assert(() {
-///       overflowRenderObject ??= this;
-///       overflowPaintIndicator(context, offset);
-///       return true;
-///     });
-///   }
-///
-///   @override
-///   void reassemble() {
-///     super.reassemble();
-///     assert(() {
-///       overflowReportNeeded = true;
-///       return true;
-///     }());
-///   }
-/// }
-/// ```
-///
-/// See also:
-///   * The code for [RenderUnconstrainedBox], and [RenderFlex] for examples of
-///     classes that use this indicator mixin.
-class DebugOverflowIndicatorMixin {
-  // This class is intended to be used as a mixin, and should not be
-  // extended directly.
-  factory DebugOverflowIndicatorMixin._() => null;
-
-  static const Color _black = const Color(0xBF000000);
-  static const Color _yellow = const Color(0xBFFFFF00);
-  static const double _kMarkerSize = 0.1;
-  static const double _debugMarkerFontSize = 7.5;
-  static const double _debugMarkerLabelPadding = 1.0;
-  static const TextStyle _debugMarkerTextStyle = const TextStyle(
-    color: const Color(0xFF900000),
-    fontSize: _debugMarkerFontSize,
-    fontWeight: FontWeight.w800,
-  );
-  static final Paint _debugMarkerPaint = new Paint()
-    ..shader = new ui.Gradient.linear(
-      const Offset(0.0, 0.0),
-      const Offset(10.0, 10.0),
-      <Color>[_black, _yellow, _yellow, _black],
-      <double>[0.25, 0.25, 0.75, 0.75],
-      TileMode.repeated,
-    );
-  static final Paint _debugLabelBackgroundPaint = new Paint()..color = const Color(0xFFFFFFFF);
-
-  final TextPainter _debugMarkerLabel = new TextPainter()
-    ..textDirection = TextDirection.ltr; // This label is in English.
-
-  TextDirection get textDirection => null;
-
-  RenderObject overflowRenderObject;
-
-  String get _containerName => runtimeType.toString();
-
-  /// Set to true to trigger a debug message in the console upon
-  /// the next paint call.
-  ///
-  /// Will be set to false after a call to
-  /// paint, so it must be reset each time a debug message is desired.
-  ///
-  /// Typically, this is set to true inside of the
-  /// [RenderObject.reassemble] override in the [RenderObject] that
-  /// owns this [DebugOverflowIndicatorMixin].
-  bool overflowReportNeeded = true;
-
-  /// An explanation as to what went wrong, and how to fix it.
-  ///
-  /// Override this to customize the hints given when an overflow occurs.  The
-  /// amount of the overflow in each direction, and the details of the
-  /// [RenderObject] that is overflowing will be printed along with this
-  /// message.
-  String get overflowHints {
-    return 'The edge of the $runtimeType that is overflowing has been '
-        'marked in the rendering with a yellow and black striped pattern. This '
-        'is usually caused by the contents being too big for the '
-        '$runtimeType.\n'
-        'This is considered an error condition because it indicates that there '
-        'is content that cannot be seen. If the content is legitimately bigger '
-        'than the available space, consider clipping it with a RectClip widget '
-        'before putting it in the $runtimeType, or using a scrollable '
-        'container, like a ListView, for example.';
-  }
-
-  /// Returns a [RelativeRect] that describes the overflow conditions
-  /// calculated.
-  ///
-  /// Values in the result will be positive if there is an overflow on
-  /// that side.
-  RelativeRect get overflow {
-    _overflow ??= new RelativeRect.fromRect(
-      _overflowContainerRect,
-      _overflowChildRect,
-    );
-    return _overflow;
-  }
-
-  RelativeRect _overflow;
-
-  /// The [Rect] of the child of the container to be checked for
-  /// overflow.
-  ///
-  /// Both [overflowChildRect] and [overflowContainerRect] must be set before
-  /// the overflow will be calculated correctly.
-  Rect get overflowChildRect => _overflowChildRect;
-  set overflowChildRect(Rect overflowChildRect) {
-    if (overflowChildRect == _overflowChildRect)
-      return;
-    _overflowChildRect = overflowChildRect;
-    _markOverflowNeedsUpdate();
-  }
-
-  Rect _overflowChildRect = Rect.zero;
-
-  /// The [Rect] of the container that is to be checked for overflow.
-  ///
-  /// Both [overflowChildRect] and [overflowContainerRect] must be set before
-  /// the overflow will be calculated correctly.
-  Rect get overflowContainerRect => _overflowContainerRect;
-  set overflowContainerRect(Rect overflowContainerRect) {
-    if (overflowContainerRect == _overflowContainerRect)
-      return;
-    _overflowContainerRect = overflowContainerRect;
-    _markOverflowNeedsUpdate();
-  }
-
-  Rect _overflowContainerRect = Rect.zero;
-
-  void _markOverflowNeedsUpdate() {
-    _overflow = null;
-  }
-
-  /// Returns true if the [overflowContainerRect] has been found to overflow.
-  bool get isOverflowing =>
-      overflow.left > 0.0 ||
-      overflow.right > 0.0 ||
-      overflow.top > 0.0 ||
-      overflow.bottom > 0.0;
-
-  String _formatPixels(double value) {
-    String pixels = 'zero';
-    if (value <= 0.0)
-      return pixels;
-    if (value > 10.0) {
-      pixels = value.toStringAsFixed(0);
-    } else if (value > 1.0) {
-      pixels = value.toStringAsFixed(1);
-    } else {
-      pixels = value.toStringAsPrecision(3);
-    }
-    return pixels;
-  }
-
-  List<_OverflowRegionData> _calculateOverflowRegions() {
-    final List<_OverflowRegionData> regions = <_OverflowRegionData>[];
-    if (overflow.left > 0.0) {
-      final Rect markerRect = new Rect.fromLTWH(
-        0.0,
-        0.0,
-        _overflowContainerRect.width * _kMarkerSize,
-        _overflowContainerRect.height,
-      );
-      String label;
-      final String pixels = _formatPixels(overflow.left);
-      if (textDirection != null) {
-        switch (textDirection) {
-          case TextDirection.ltr:
-            label = 'BEGIN OVERFLOWED BY $pixels PIXELS';
-            break;
-          case TextDirection.rtl:
-            label = 'END OVERFLOWED BY $pixels PIXELS';
-            break;
-        }
-      } else {
-        label = 'LEFT OVERFLOWED BY $pixels PIXELS';
-      }
-      regions.add(new _OverflowRegionData(
-        rect: markerRect,
-        label: label,
-        labelOffset: markerRect.centerLeft +
-            const Offset(_debugMarkerFontSize + _debugMarkerLabelPadding, 0.0),
-        rotation: math.PI / 2.0,
-      ));
-    }
-    if (overflow.right > 0.0) {
-      final Rect markerRect = new Rect.fromLTWH(
-        _overflowContainerRect.width * (1.0 - _kMarkerSize),
-        0.0,
-        _overflowContainerRect.width * _kMarkerSize,
-        _overflowContainerRect.height,
-      );
-      String label;
-      final String pixels = _formatPixels(overflow.right);
-      if (textDirection != null) {
-        switch (textDirection) {
-          case TextDirection.ltr:
-            label = 'END OVERFLOWED BY $pixels PIXELS';
-            break;
-          case TextDirection.rtl:
-            label = 'BEGIN OVERFLOWED BY $pixels PIXELS';
-            break;
-        }
-      } else {
-        label = 'RIGHT OVERFLOWED BY $pixels PIXELS';
-      }
-      regions.add(new _OverflowRegionData(
-        rect: markerRect,
-        label: label,
-        labelOffset: markerRect.centerRight -
-            const Offset(_debugMarkerFontSize + _debugMarkerLabelPadding, 0.0),
-        rotation: -math.PI / 2.0,
-      ));
-    }
-    if (overflow.top > 0.0) {
-      final String pixels = _formatPixels(overflow.top);
-      final Rect markerRect = new Rect.fromLTWH(
-        0.0,
-        0.0,
-        _overflowContainerRect.width,
-        _overflowContainerRect.height * _kMarkerSize,
-      );
-      regions.add(new _OverflowRegionData(
-        rect: markerRect,
-        label: 'TOP OVERFLOWED BY $pixels PIXELS',
-        labelOffset: markerRect.topCenter + const Offset(0.0, _debugMarkerLabelPadding),
-        rotation: 0.0,
-      ));
-    }
-    if (overflow.bottom > 0.0) {
-      final String pixels = _formatPixels(overflow.bottom);
-      final Rect markerRect = new Rect.fromLTWH(
-        0.0,
-        _overflowContainerRect.height * (1.0 - _kMarkerSize),
-        _overflowContainerRect.width,
-        _overflowContainerRect.height * _kMarkerSize,
-      );
-      regions.add(new _OverflowRegionData(
-        rect: markerRect,
-        label: 'BOTTOM OVERFLOWED BY $pixels PIXELS',
-        labelOffset: markerRect.bottomCenter -
-            const Offset(0.0, _debugMarkerFontSize + _debugMarkerLabelPadding),
-        rotation: 0.0,
-      ));
-    }
-    return regions;
-  }
-
-  void _reportOverflow() {
-    final List<String> overflows = <String>[];
-    if (overflow.left > 0.0)
-      overflows.add(_formatPixels(overflow.left) + ' pixels on the left');
-    if (overflow.top > 0.0)
-      overflows.add(_formatPixels(overflow.top) + ' pixels on the top');
-    if (overflow.bottom > 0.0)
-      overflows.add(_formatPixels(overflow.bottom) + ' pixels on the bottom');
-    if (overflow.right > 0.0)
-      overflows.add(_formatPixels(overflow.right) + ' pixels on the right');
-    String overflowText = '';
-    assert(overflows.isNotEmpty,
-        "Somehow $_containerName didn't actually overflow like it thought it did.");
-    switch (overflows.length) {
-      case 1:
-        overflowText = overflows.first;
-        break;
-      case 2:
-        overflowText = '${overflows.first} and ${overflows.last}';
-        break;
-      default:
-        overflows[overflows.length - 1] = 'and ${overflows[overflows.length - 1]}';
-        overflowText = overflows.join(', ');
-    }
-    FlutterError.reportError(
-      new FlutterErrorDetailsForRendering(
-        exception: 'A $_containerName overflowed by $overflowText.',
-        library: 'rendering library',
-        context: 'during layout',
-        renderObject: overflowRenderObject,
-        informationCollector: (StringBuffer information) {
-          information.writeln(overflowHints);
-          information.writeln('The specific $_containerName in question is:');
-          information.writeln('  ${overflowRenderObject.toStringShallow(joiner: '\n  ')}');
-          information.writeln('◢◤' * (FlutterError.wrapWidth ~/ 2));
-        },
-      ),
-    );
-  }
-
-  /// To be called when the overflow indicators should be painted.
-  ///
-  /// Typically only called if there is an overflow, and only from within a
-  /// debug build.
-  ///
-  /// See example code in [DebugOverflowIndicatorMixin] documentation.
-  void overflowPaintIndicator(PaintingContext context, Offset offset) {
-    assert(overflowRenderObject != null,
-        'Must set overflowRenderObject before calling overflowPaintIndicator',);
-    if (!isOverflowing) {
-      return;
-    }
-
-    final List<_OverflowRegionData> overflowRegions = _calculateOverflowRegions();
-    for (_OverflowRegionData region in overflowRegions) {
-      context.canvas.drawRect(region.rect.shift(offset), _debugMarkerPaint);
-
-      _debugMarkerLabel.text = new TextSpan(
-        // This is a no-op if the label hasn't changed.
-        text: region.label,
-        style: _debugMarkerTextStyle,
-      );
-      _debugMarkerLabel.layout(); // This is a no-op if the label hasn't changed.
-
-      final Offset labelOffset = region.labelOffset + offset;
-      final Offset centerOffset = new Offset(-_debugMarkerLabel.width / 2.0, 0.0);
-      final Rect textBackgroundRect = centerOffset & _debugMarkerLabel.size;
-      context.canvas.save();
-      context.canvas.translate(labelOffset.dx, labelOffset.dy);
-      context.canvas.rotate(region.rotation);
-      context.canvas.drawRect(textBackgroundRect, _debugLabelBackgroundPaint);
-      _debugMarkerLabel.paint(context.canvas, centerOffset);
-      context.canvas.restore();
-    }
-
-    if (overflowReportNeeded) {
-      overflowReportNeeded = false;
-      _reportOverflow();
-    }
-  }
-}
